+++ packages/rockset-node-client/src/rest/model/CreateIntegrationRequest.js
@@ -1,18 +1,18 @@
 (function(root, factory) {
   if (typeof define === 'function' && define.amd) {
     // AMD. Register as an anonymous module.
-    define(['ApiClient', 'model/AwsKeyIntegration'], factory);
+    define(['ApiClient', 'model/AwsExternalIdIntegration', 'model/AwsKeyIntegration', 'model/GcpServiceAccount'], factory);
   } else if (typeof module === 'object' && module.exports) {
     // CommonJS-like environments that support module.exports, like Node.
-    module.exports = factory(require('../ApiClient'), require('./AwsKeyIntegration'));
+    module.exports = factory(require('../ApiClient'), require('./AwsExternalIdIntegration'), require('./AwsKeyIntegration'), require('./GcpServiceAccount'));
   } else {
     // Browser globals (root is window)
     if (!root.RestApi) {
       root.RestApi = {};
     }
-    root.RestApi.CreateIntegrationRequest = factory(root.RestApi.ApiClient, root.RestApi.AwsKeyIntegration);
+    root.RestApi.CreateIntegrationRequest = factory(root.RestApi.ApiClient, root.RestApi.AwsExternalIdIntegration, root.RestApi.AwsKeyIntegration, root.RestApi.GcpServiceAccount);
   }
-}(this, function(ApiClient, AwsKeyIntegration) {
+}(this, function(ApiClient, AwsExternalIdIntegration, AwsKeyIntegration, GcpServiceAccount) {
     'use strict';
 
 
@@ -34,6 +34,8 @@
     var _this = this;
 
     _this['name'] = name;
+
+
 
 
   };
@@ -58,6 +60,12 @@
       if (data.hasOwnProperty('aws')) {
         obj['aws'] = AwsKeyIntegration.constructFromObject(data['aws']);
       }
+      if (data.hasOwnProperty('aws_external_id')) {
+        obj['aws_external_id'] = AwsExternalIdIntegration.constructFromObject(data['aws_external_id']);
+      }
+      if (data.hasOwnProperty('gcp_service_account')) {
+        obj['gcp_service_account'] = GcpServiceAccount.constructFromObject(data['gcp_service_account']);
+      }
     }
     return obj;
   }
@@ -77,6 +85,16 @@
    * @member {module:model/AwsKeyIntegration} aws
    */
   exports.prototype['aws'] = undefined;
+  /**
+   * details for an AWS External Id integration
+   * @member {module:model/AwsExternalIdIntegration} aws_external_id
+   */
+  exports.prototype['aws_external_id'] = undefined;
+  /**
+   * details of a GCP Service Account integration
+   * @member {module:model/GcpServiceAccount} gcp_service_account
+   */
+  exports.prototype['gcp_service_account'] = undefined;